added logger

+++ rpm_reader.py
@@ -15,7 +15,7 @@ def get_rpm():
             return 0
         if ser.in_waiting > 0:
             try:
-                rpm = float(ser.readline().decode('utf-8').rstrip())
+                rpm = int(ser.readline().decode('utf-8').rstrip())
                 print(type(rpm))
                 return rpm
             except UnicodeDecodeError as e:
